backend/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
sqlalchemy
pgvector
PyJWT>=2.9.0
